Transaction counter validation

+++ src/test/java/de/konfidas/ttc/setup/Utilities.java
@@ -7,13 +7,11 @@ import org.bouncycastle.util.encoders.Base64;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.PrintWriter;
-import java.math.BigInteger;
 import java.nio.ByteBuffer;
 import java.security.KeyPair;
 import java.security.KeyStore;
 import java.security.cert.Certificate;
 import java.util.Arrays;
-import java.util.Random;
 
 public class Utilities {
 
@@ -42,14 +40,14 @@ public class Utilities {
     public static int getEncodedLength(ASN1Primitive element) throws IOException, ExtendLengthValueExceedsInteger {
         byte[] elementContent = element.getEncoded();
 
-        if ((byte) elementContent[1] == (byte) 0b10000000) {
+        if (elementContent[1] == (byte) 0b10000000) {
             //indefinte length encoding
             return 0;
         }
 
         else if ((elementContent[1] & 0b10000000) == 0) {
             //Case: Definite length encocding, one byte
-            return Integer.valueOf(elementContent[1]);
+            return elementContent[1];
         }
 
         else {
